second stage cleanup
