+++ bot/ctftime.py
@@ -1,22 +1,18 @@
-import requests
-import re
-from urllib.request import urlopen
 import io
-from datetime import *
+import re
+from datetime import datetime, timezone
+
+from urllib.request import urlopen
 from dateutil.parser import parse
 
-from colorthief import ColorThief
+import requests
 import discord
 from discord.ext import commands
-from vars.help_info import (
-    embed_help,
-    ctf_help_text,
-    chal_help_text,
-)
-from controllers.db import ctfs, teamdb
-import models.ctf as ctfmodel
+from colorthief import ColorThief
 from lxml import html
 
+import db
+from config import config
 
 
 class Ctftime(commands.Cog):
@@ -31,8 +27,8 @@ class Ctftime(commands.Cog):
     default_image = "https://pbs.twimg.com/profile_images/2189766987/ctftime-logo-avatar_400x400.png"
 
     @staticmethod
-    def rgb2hex(r, g, b):
-        tohex = "#{:02x}{:02x}{:02x}".format(r, g, b)
+    def rgb2hex(response, g, b):
+        tohex = "#{:02x}{:02x}{:02x}".format(response, g, b)
         return tohex
 
     @staticmethod
@@ -81,11 +77,11 @@ class Ctftime(commands.Cog):
         for ctf in info:  # If the document doesn't exist: add it, if it does: update it.
             print(f"Got {ctf['name']} from ctftime")
             query = ctf["name"]
-            ctfs.update({"name": query}, {"$set": ctf}, upsert=True)
+            db.ctfs.update({"name": query}, {"$set": ctf}, upsert=True)
 
-        for ctf in ctfs.find():  # Delete ctfs that are over from the db
+        for ctf in db.ctfs.find():  # Delete ctfs that are over from the db
             if ctf["end"] < unix_now:
-                ctfs.remove({"name": ctf["name"]})
+                db.ctfs.remove({"name": ctf["name"]})
 
     @commands.group()
     async def ctftime(self, ctx):
@@ -126,10 +122,6 @@ class Ctftime(commands.Cog):
             else:
                 ctf_place = "Onsite"
 
-            # if ctf_image != '':
-            #     fd = urlopen(ctf_image) # 403 is on this line, no longer able to access this?
-            # else:
-            #     fd = urlopen(Ctftime.default_image)
             fd = urlopen(Ctftime.default_image)
             f = io.BytesIO(fd.read())
             color_thief = ColorThief(f)
@@ -184,30 +176,29 @@ class Ctftime(commands.Cog):
             f""":triangular_flag_on_post:  **{params} CTFtime Leaderboards**```ini
 {leaderboards}```"""
         )
-    
+
     @ctftime.command()
     async def team(self, ctx, team=None):
-        team_id = 119480 # EPT
+        team_id = config["team"]["id"]
         if team:
             msg = await ctx.send(f"Looking id for team {team}...")
             team_id = get_team_id(team)
             await msg.delete()
         else:
-            team = 'EPT'
+            team = config["team"]["name"]
         if team_id <= 0:
             ctx.send(f':warning: Unknown team `{team}`.')
             return
         msg = await ctx.send(f"Looking up scores for {team} with id {team_id}...")
         table = get_scores(team_id)
         await msg.delete()
-        table = [line[:2] + line[3:] for line in table] # remove CTF points column, not interesting
-        table = table[:11] # top 10 scores, already sorted by rating points
+        table = [line[:2] + line[3:] for line in table]  # remove CTF points column, not interesting
+        table = table[:11]  # top 10 scores, already sorted by rating points
         out = f'Top 10 events for {team}'
         out += '```'
         out += format_table(table[:10])
         out += '```'
         await ctx.send(out)
-
 
     @ctftime.command()
     async def current(self, ctx, params=None):
@@ -216,10 +207,8 @@ class Ctftime(commands.Cog):
         unix_now = int(now.replace(tzinfo=timezone.utc).timestamp())
         running = False
 
-        for ctf in ctfs.find():
-            if (
-                ctf["start"] < unix_now and ctf["end"] > unix_now
-            ):  # Check if the ctf is running
+        for ctf in db.ctfs.find():
+            if (ctf["start"] < unix_now and unix_now < ctf["end"]):  # Check if the ctf is running
                 running = True
                 embed = discord.Embed(
                     title=":red_circle: " + ctf["name"] + " IS LIVE",
@@ -258,7 +247,7 @@ class Ctftime(commands.Cog):
         now = datetime.utcnow()
         unix_now = int(now.replace(tzinfo=timezone.utc).timestamp())
         running = False
-        for ctf in ctfs.find():
+        for ctf in db.ctfs.find():
             if (
                 ctf["start"] < unix_now and ctf["end"] > unix_now
             ):  # Check if the ctf is running
@@ -275,7 +264,7 @@ class Ctftime(commands.Cog):
                     f"```ini\n{ctf['name']} ends in: [{days} days], [{hours} hours], [{minutes} minutes], [{seconds} seconds]```\n{ctf['url']}"
                 )
 
-        if running == False:
+        if not running:
             await ctx.send(
                 "No ctfs are running! Use !ctftime upcoming or !ctftime countdown to see upcoming ctfs"
             )
@@ -286,14 +275,14 @@ class Ctftime(commands.Cog):
         now = datetime.utcnow()
         unix_now = int(now.replace(tzinfo=timezone.utc).timestamp())
 
-        if params == None:
+        if params is None:
             self.upcoming_l = []
             index = ""
-            for ctf in ctfs.find():
+            for ctf in db.ctfs.find():
                 if ctf["start"] > unix_now:
                     self.upcoming_l.append(ctf)
-            for i, c in enumerate(self.upcoming_l):
-                index += f"\n[{i + 1}] {c['name']}\n"
+            for i, ctf in enumerate(self.upcoming_l):
+                index += f"\n[{i + 1}] {ctf['name']}\n"
 
             await ctx.send(
                 f"Type !ctftime countdown <number> to select.\n```ini\n{index}```"
@@ -302,18 +291,6 @@ class Ctftime(commands.Cog):
         else:
             if self.upcoming_l != []:
                 x = int(params) - 1
-                start = (
-                    datetime.utcfromtimestamp(self.upcoming_l[x]["start"]).strftime(
-                        "%Y-%m-%d %H:%M:%S"
-                    )
-                    + " UTC"
-                )
-                end = (
-                    datetime.utcfromtimestamp(self.upcoming_l[x]["end"]).strftime(
-                        "%Y-%m-%d %H:%M:%S"
-                    )
-                    + " UTC"
-                )
 
                 time = self.upcoming_l[x]["start"] - unix_now
                 days = time // (24 * 3600)
@@ -328,22 +305,10 @@ class Ctftime(commands.Cog):
                     f"```ini\n{self.upcoming_l[x]['name']} starts in: [{days} days], [{hours} hours], [{minutes} minutes], [{seconds} seconds]```\n{self.upcoming_l[x]['url']}"
                 )
             else:  # TODO: make this a function, too much repeated code here.
-                for ctf in ctfs.find():
+                for ctf in db.ctfs.find():
                     if ctf["start"] > unix_now:
                         self.upcoming_l.append(ctf)
                 x = int(params) - 1
-                start = (
-                    datetime.utcfromtimestamp(self.upcoming_l[x]["start"]).strftime(
-                        "%Y-%m-%d %H:%M:%S"
-                    )
-                    + " UTC"
-                )
-                end = (
-                    datetime.utcfromtimestamp(self.upcoming_l[x]["end"]).strftime(
-                        "%Y-%m-%d %H:%M:%S"
-                    )
-                    + " UTC"
-                )
 
                 time = self.upcoming_l[x]["start"] - unix_now
                 days = time // (24 * 3600)
@@ -357,6 +322,7 @@ class Ctftime(commands.Cog):
                 await ctx.send(
                     f"```ini\n{self.upcoming_l[x]['name']} starts in: [{days} days], [{hours} hours], [{minutes} minutes], [{seconds} seconds]```\n{self.upcoming_l[x]['url']}"
                 )
+
 
 def get_scores(team_id):
     url = f"https://ctftime.org/team/{team_id}"
@@ -376,266 +342,28 @@ def get_scores(team_id):
     table = [column_names] + table
     return table
 
+
 def get_team_id(team_name):
     ses = requests.Session()
     url = 'https://ctftime.org/stats/'
-    r = ses.get(url, headers=Ctftime.headers)
-    doc = html.fromstring(r.content)
+    response = ses.get(url, headers=Ctftime.headers)
+    doc = html.fromstring(response.content)
     token = doc.xpath('//input[@name="csrfmiddlewaretoken"]')[0]
 
     url = "https://ctftime.org/team/list/"
-    h = {'Referer': 'https://ctftime.org/stats/'}
-    h.update(Ctftime.headers)
-    r = ses.post(url, data={'team_name': team_name, 'csrfmiddlewaretoken': token.value}, headers=h)
-    team_id = r.url.split("/")[-1]
+    headers = {'Referer': 'https://ctftime.org/stats/'}
+    headers.update(Ctftime.headers)
+    response = ses.post(url, data={'team_name': team_name, 'csrfmiddlewaretoken': token.value}, headers=headers)
+    team_id = response.url.split("/")[-1]
     if team_id.isnumeric():
         return int(team_id)
     return -1
+
 
 def format_table(table, seperator='      '):
     widths = [max(len(line[i]) for line in table) for i in range(len(table[0]))]
     return '\n'.join([seperator.join([c.ljust(w) for w, c in zip(widths, line)]) for line in table])
 
-async def respond(ctx, fn, *args):
-    messages = []
-    guild = ctx.channel.guild
-    async with ctx.channel.typing():
-        for chan_id, msg in await fn(*args):
-            chan = guild.get_channel(chan_id) if chan_id else ctx.channel
-            msg = await chan.send(msg)
-            messages.append(msg)
-    return messages
-
-async def respond_with_reaction(ctx, emoji, fn, *args):
-    messages = []
-    guild = ctx.channel.guild
-    async with ctx.channel.typing():
-        for chan_id, msg in await fn(*args):
-            chan = guild.get_channel(chan_id) if chan_id else ctx.channel
-            msg = await chan.send(msg)
-            await msg.add_reaction(emoji)
-            messages.append(msg)
-    return messages
-
-
-def check_name(name):
-    if len(name) > 32:
-        raise ctfmodel.TaskFailed("Challenge name is too long!")
-
-    if not re.match(r"[-.!0-9A-Za-z ]+$", name):
-        raise ctfmodel.TaskFailed("Challenge contains invalid characters!")
-
-    # Replace spaces with a dash, because discord does it :/
-    return re.sub(r" +", "-", name).lower()
-
-
-def chk_fetch_team_by_name(ctx, name):
-    channels = ctx.guild.channels
-    if len([channel.id for channel in channels if name == channel.name]) > 1:
-        raise ctfmodel.TaskFailed("Multiple channels with same name exists")
-
-    found_channel_id = ""
-    for channel in channels:
-        if channel.name == name:
-            found_channel_id = channel.id
-    team = ctfmodel.CtfTeam.fetch(ctx.channel.guild, found_channel_id)
-    if not team:
-        raise ctfmodel.TaskFailed("Failed to join CTF")
-    return team
-
-
-def chk_fetch_team(ctx):
-    team = ctfmodel.CtfTeam.fetch(ctx.channel.guild, ctx.channel.id)
-    if not team:
-        raise ctfmodel.TaskFailed("Please type this command in a ctf channel.")
-    return team
-
-
-def chk_fetch_chal(ctx):
-    chal = ctfmodel.Challenge.fetch(ctx.channel.guild, ctx.channel.id)
-    if not chal:
-        raise ctfmodel.TaskFailed(
-            "Please type this command in a challenge "
-            + "channel. You may need to join a challenge first."
-        )
-    return chal
-
-
-def parse_user(guild, user):
-    mat = re.match(r"<@([0-9]+)>$", user)
-    if mat:
-        ret = guild.get_member(int(mat[1]))
-    else:
-        ret = guild.get_member_named(user)
-
-    if not ret:
-        raise ctfmodel.TaskFailed(f'Invalid username: "{user}"')
-    return ret
-
-
-def verify_owner():
-    def predicate(ctx):
-        chk_fetch_chal(ctx).check_done(ctx.author)
-        return True
-
-    return commands.check(predicate)
-
-
-class Ctfs(commands.Cog):
-    def __init__(self, bot):
-        self.bot = bot
-        self.challenges = {}
-        self.ctfname = ""
-
-    @commands.bot_has_permissions(manage_roles=True, manage_channels=True)
-    @commands.guild_only()
-    @commands.command()
-    async def create(self, ctx, name):
-        emoji = '🏃'
-        messages = await respond_with_reaction(ctx, emoji, ctfmodel.CtfTeam.create, ctx.channel.guild, name)
-        teamdb[str(ctx.channel.guild.id)].update_one(
-            {"name": name}, {"$set": {"msg_id": messages[0].id}}
-        )
-
-    @commands.guild_only()
-    @commands.group()
-    async def ctf(self, ctx):
-        if ctx.invoked_subcommand is None:
-            await ctx.send("Invalid command passed. Use !ctf help.")
-
-    @ctf.command("help")
-    async def ctf_help(self, ctx):
-        await embed_help(ctx, "CTF team help topic", ctf_help_text)
-
-    @commands.bot_has_permissions(manage_channels=True)
-    @ctf.command()
-    async def add(self, ctx, name):
-        name = check_name(name)
-        emoji = '🔨'
-        await respond_with_reaction(ctx, emoji,  chk_fetch_team(ctx).add_chal, name)
-
-    @commands.bot_has_permissions(manage_channels=True)
-    @commands.has_permissions(manage_channels=True)
-    @ctf.command()
-    async def delete(self, ctx, name):
-        name = check_name(name)
-        await respond(ctx, chk_fetch_team(ctx).del_chal, name)
-
-    @commands.bot_has_permissions(manage_roles=True)
-    @commands.command("leave")
-    async def leave_ctf(self, ctx):
-        await respond(ctx, chk_fetch_team(ctx).leave, ctx.author)
-
-    @commands.bot_has_permissions(manage_roles=True)
-    @ctf.command("invite")
-    async def invite_ctf(self, ctx, user):
-        user = parse_user(ctx.channel.guild, user)
-        await respond(ctx, chk_fetch_team(ctx).invite, ctx.author, user)
-
-    @commands.bot_has_permissions(manage_roles=True)
-    @commands.command()
-    async def join(self, ctx, name):
-        await respond(ctx, chk_fetch_team_by_name(ctx, name).join, ctx.author)
-
-    @ctf.command()
-    async def working(self, ctx, chalname):
-        chk_fetch_team(ctx)
-        chal = ctfmodel.Challenge.find(ctx.channel.guild, ctx.channel.id, chalname)
-        await respond(ctx, chal.working, ctx.author)
-
-    @commands.bot_has_permissions(manage_roles=True, manage_channels=True)
-    @commands.has_permissions(manage_channels=True)
-    @ctf.command()
-    async def archive(self, ctx):
-        await respond(ctx, chk_fetch_team(ctx).archive)
-
-    @commands.bot_has_permissions(manage_roles=True, manage_channels=True)
-    @commands.has_permissions(manage_channels=True)
-    @ctf.command()
-    async def unarchive(self, ctx):
-        await respond(ctx, chk_fetch_team(ctx).unarchive)
-
-    @commands.bot_has_permissions(manage_roles=True, manage_channels=True)
-    @commands.has_permissions(manage_channels=True)
-    @ctf.command()
-    async def export(self, ctx):
-        await respond(ctx, ctfmodel.export, ctx, ctx.author)
-
-    @commands.bot_has_permissions(manage_roles=True, manage_channels=True)
-    @commands.has_permissions(manage_channels=True)
-    @ctf.command()
-    async def deletectf(self, ctx):
-        await respond(ctx, ctfmodel.delete, ctx, ctx.author)
-
-    @ctf.command()
-    async def list(self, ctx):
-        chals = chk_fetch_team(ctx).challenges
-        if len(chals) == 0:
-            await ctx.send("No challenges added...")
-            return
-
-        msg_len = 50
-        lines = []
-        for chal in chals:
-            l = f"[{chal.team.name}] [{chal.name}] - {chal.status}"
-            msg_len += len(l) + 1
-            if msg_len > 1000:  # Over limit
-                lines = "\n".join(lines)
-                await ctx.send(f"```ini\n{lines}```")
-                lines = []
-                msg_len = len(l) + 51
-            lines.append(l)
-
-        lines = "\n".join(lines)
-        await ctx.send(f"```ini\n{lines}```")
-
-    @commands.guild_only()
-    @commands.group()
-    async def chal(self, ctx):
-        if ctx.invoked_subcommand is None:
-            await ctx.send("Invalid command passed.  Use !help.")
-
-    @commands.bot_has_permissions(manage_channels=True)
-    @chal.command("invite")
-    async def invite_chal(self, ctx, user):
-        user = parse_user(ctx.channel.guild, user)
-        await respond(ctx, chk_fetch_chal(ctx).invite, ctx.author, user)
-
-    @commands.bot_has_permissions(manage_channels=True)
-    @verify_owner()
-    @chal.command()
-    async def done(self, ctx, *withlist):
-        guild = ctx.channel.guild
-        users = [parse_user(guild, u) for u in withlist]
-        await respond(ctx, chk_fetch_chal(ctx).done, ctx.author, users)
-
-    @chal.command("help")
-    async def chal_help(self, ctx):
-        await embed_help(ctx, "Challenge help topic", chal_help_text)
-
-    @commands.bot_has_permissions(manage_channels=True)
-    @verify_owner()
-    @chal.command()
-    async def undone(self, ctx):
-        await respond(ctx, chk_fetch_chal(ctx).undone)
-
-    @commands.bot_has_permissions(manage_channels=True)
-    @chal.command("leave")
-    async def leave_chal(self, ctx):
-        await respond(ctx, chk_fetch_chal(ctx).leave, ctx.author)
-
-    @commands.command()
-    async def htb(self, ctx):
-        twitter_page = requests.get("https://twitter.com/hackthebox_eu")
-        all_content = str(twitter_page.text.encode("utf-8"))
-        tweet = re.search(
-            "\\w+ will go live \\d{2}/\\d{2}/\\d{4} at \\d{2}:\\d{2}:\\d{2} UTC",
-            all_content,
-        )
-        match = tweet.group(0)
-        await ctx.send(match + "\nhttps://hackthebox.eu")
-
 
 def setup(bot):
-    bot.add_cog(Ctfs(bot))
     bot.add_cog(Ctftime(bot))
