+++ bot/help_info.py
@@ -41,47 +41,6 @@ return the latest tweet from @hackthebox_eu that says when the next box will be 
 """
 
 help_page_2 = """
-
-
-`!rot <message> <direction(optional, will default to left)>`
-return all 25 different possible combinations for the popular caesar cipher - use quotes for messages more than 1 word
-
-`!magicb <filetype>`
-return the magicbytes/file header of a supplied filetype.
-
-`!b64 <encode/decode> <message>`
-encode or decode in base64 - if message has spaces use quotations
-
-`!binary <encode/decode> <message>`
-encode or decode in binary - if message has spaces use quotations
-
-`!hex <encode/decode> <message>`
-encode or decode in hex - if message has spaces use quotations
-
-`!url <encode/decode> <message>`
-encode or decode based on url encoding - if message has spaces use quotations
-
-`!reverse <message>`
-reverse the supplied string - if message has spaces use quotations
-
-`!counteach <message>`
-count the occurrences of each character in the supplied message - if message has spaces use quotations
-
-`!characters <message>`
-count the amount of characters in your supplied message
-
-`!wordcount <phrase>`
-count the amount of words in your supplied message
-
-`!atbash <message>`
-encode or decode in the atbash cipher - if message has spaces use quotations (encode/decode do the same thing)
-
-`!github <user>`
-get a direct link to a github profile page with your supplied user
-
-`!twitter <user>`
-get a direct link to a twitter profile page with your supplied user
-
 `!cointoss`
 get a 50/50 cointoss to make all your life's decisions
