+++ issues-todo
@@ -1,0 +1,1 @@
+* fix "!chal done <username>" to accept @ in usernames
