--- vars/general.py
@@ -1,2 +1,0 @@
-
-cool_names = ["null", "nordbo", "LOLASL", "rckwl", "vcpo"]
